Added jump anim + trail renderer
The player feels how I wanted them to feel, I need to work on the dash tomorrow, hopefully get that working in the basic form, then make sure it feels exactly how I imagined it feeling. After that I need to start either on enemies or a demo scene that I can send to people for testing.