Added Particles for Jumping and running.
Today had ups and downs. I did a lot of work but if I am being honest its getting messy. There are a lot of references to the same thing in many places and I am starting to worry if I have to many things connected to the player. Building the game looks fine but again for some reason the editor is shitting the bed. I have no idea why but its getting annoying. On a more positive note now that the particles are done for the most part I can start on finishing the other abilities the player has. Those should be relatively simple and self contained so fingers crossed it goes easier.