Testing for Framerate and prep for abilites
I did some testing for the framerate and I found that during a build my framerate is jumping from 60-200 frames so im not worrried during a build. But! during editor the framerate is around 30. Thats annoying as hell. whatever the reason I'm gunna have to figure it out. The game is smooth in build so its time to move on to adding new abilities. Tomorrow will be Adding the blast and the Ground Pound. I did some prep for it tonight by writing out all the methods I just need to populate them. I think I have a good number of abilities that enables the player to really put the hurt on their foes. Once these abilities are all squared away and I like the feel of them I think I am ready to make the first level.