Ground Pound
Ground Pound is working, right now all that needs to be added to it is a reference to the enemy script that enables damage. other then that though its pretty good. Moving onto generic attack probably tomorrow and hopefully get the other two attacks by the end of the weekend.